requirements.txt
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
requests
gunicorn==20.1.0
